requirements.txt
boto3
requests
PyYAML
pynacl
